+++ packages/relax/src/hooks/use-controlled-state.ts
@@ -2,8 +2,8 @@ import { type Dispatch, type SetStateAction, useState } from "react";
 import type { Partialable, RequiredIn, State } from "@aiszlab/relax/types";
 import { isUndefined } from "../is/is-undefined";
 import { isFunction } from "../is/is-function";
+import { useUpdateEffect } from "./use-update-effect";
 import { useEvent } from "./use-event";
-import { flushSync } from "react-dom";
 
 type UsingControlledState<S> = {
   /**
@@ -60,30 +60,16 @@ function useControlledState<T>(
     return defaultState ?? controlledState;
   });
 
-  // 变更状态函数
-  // 变更内容为函数，且需要触发`onChange`，那么本次状态更新必为同步
-  // 其余继续保持批处理逻辑
-  const setState = useEvent<typeof _setState>((action) => {
-    if (isFunction(action)) {
-      if (onChange) {
-        flushSync(() => {
-          _setState(action);
-        });
-
-        onChange(_state);
-        return;
-      }
-
-      _setState(action);
-      return;
-    }
-
-    _setState(action);
-    onChange?.(action);
-  });
-
   // use controlled
   const state = isUndefined(controlledState) ? _state : controlledState;
+
+  // 复写状态更新
+  const setState = useEvent<typeof _setState>((_newValue) => {
+    const _newState = isFunction(_newValue) ? _newValue(state) : _newValue;
+
+    _setState(_newState);
+    onChange?.(_newState);
+  });
 
   return [state, setState];
 }